requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
markdownify==0.11.6
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.0
